# perf(ext,runtime): remove using `SafeArrayIterator` from `for-of`