--- a/ext/fetch/21_formdata.js
+++ b/ext/fetch/21_formdata.js
@@ -25,7 +25,6 @@
     MathRandom,
     ObjectPrototypeIsPrototypeOf,
     Symbol,
-    SafeArrayIterator,
     StringFromCharCode,
     StringPrototypeTrim,
     StringPrototypeSlice,
@@ -163,7 +162,9 @@
         context: "Argument 1",
       });
 
-      for (const entry of new SafeArrayIterator(this[entryList])) {
+      const entries = this[entryList];
+      for (let i = 0; i < entries.length; ++i) {
+        const entry = entries[i];
         if (entry.name === name) return entry.value;
       }
       return null;
@@ -184,7 +185,9 @@
       });
 
       const returnList = [];
-      for (const entry of new SafeArrayIterator(this[entryList])) {
+      const entries = this[entryList];
+      for (let i = 0; i < entries.length; ++i) {
+        const entry = entries[i];
         if (entry.name === name) ArrayPrototypePush(returnList, entry.value);
       }
       return returnList;
@@ -204,7 +207,9 @@
         context: "Argument 1",
       });
 
-      for (const entry of new SafeArrayIterator(this[entryList])) {
+      const entries = this[entryList];
+      for (let i = 0; i < entries.length; ++i) {
+        const entry = entries[i];
         if (entry.name === name) return true;
       }
       return false;
@@ -374,7 +379,8 @@
     #parseHeaders(headersText) {
       const headers = new Headers();
       const rawHeaders = StringPrototypeSplit(headersText, "\r\n");
-      for (const rawHeader of new SafeArrayIterator(rawHeaders)) {
+      for (let i = 0; i < rawHeaders.length; ++i) {
+        const rawHeader = rawHeaders[i];
         const sepIndex = StringPrototypeIndexOf(rawHeader, ":");
         if (sepIndex < 0) {
           continue; // Skip this header