--- a/ext/web/01_mimesniff.js
+++ b/ext/web/01_mimesniff.js
@@ -16,7 +16,6 @@
     MapPrototypeHas,
     MapPrototypeSet,
     RegExpPrototypeTest,
-    SafeArrayIterator,
     SafeMapIterator,
     StringPrototypeReplaceAll,
     StringPrototypeToLowerCase,
@@ -223,7 +222,8 @@
     let charset = null;
     let essence_ = null;
     let mimeType = null;
-    for (const value of new SafeArrayIterator(headerValues)) {
+    for (let i = 0; i < headerValues.length; ++i) {
+      const value = headerValues[i];
       const temporaryMimeType = parseMimeType(value);
       if (
         temporaryMimeType === null ||