--- a/ext/web/13_message_port.js
+++ b/ext/web/13_message_port.js
@@ -22,7 +22,6 @@
     ArrayPrototypePush,
     ObjectPrototypeIsPrototypeOf,
     ObjectSetPrototypeOf,
-    SafeArrayIterator,
     Symbol,
     SymbolFor,
     SymbolIterator,
@@ -205,9 +204,8 @@
     const arrayBufferIdsInTransferables = [];
     const transferredArrayBuffers = [];
 
-    for (
-      const transferable of new SafeArrayIterator(messageData.transferables)
-    ) {
+    for (let i = 0; i < messageData.transferables.length; ++i) {
+      const transferable = messageData.transferables[i];
       switch (transferable.kind) {
         case "messagePort": {
           const port = createMessagePort(transferable.data);
@@ -217,8 +215,8 @@
         }
         case "arrayBuffer": {
           ArrayPrototypePush(transferredArrayBuffers, transferable.data);
-          const i = ArrayPrototypePush(transferables, null);
-          ArrayPrototypePush(arrayBufferIdsInTransferables, i);
+          const index = ArrayPrototypePush(transferables, null);
+          ArrayPrototypePush(arrayBufferIdsInTransferables, index);
           break;
         }
         default:
@@ -274,7 +272,8 @@
     const serializedTransferables = [];
 
     let arrayBufferI = 0;
-    for (const transferable of new SafeArrayIterator(transferables)) {
+    for (let i = 0; i < transferables.length; ++i) {
+      const transferable = transferables[i];
       if (ObjectPrototypeIsPrototypeOf(MessagePortPrototype, transferable)) {
         webidl.assertBranded(transferable, MessagePortPrototype);
         const id = transferable[_id];