--- a/runtime/js/99_main.js
+++ b/runtime/js/99_main.js
@@ -33,7 +33,6 @@ delete Intl.v8BreakIterator;
     SymbolFor,
     SymbolIterator,
     PromisePrototypeThen,
-    SafeArrayIterator,
     SafeWeakMap,
     TypeError,
     WeakMapPrototypeDelete,
@@ -206,7 +205,8 @@ delete Intl.v8BreakIterator;
     );
     loadedMainWorkerScript = true;
 
-    for (const { url, script } of new SafeArrayIterator(scripts)) {
+    for (let i = 0; i < scripts.length; ++i) {
+      const { url, script } = scripts[i];
       const err = core.evalContext(script, url)[1];
       if (err !== null) {
         throw err.thrown;