<!--
Before submitting a PR, please read http://deno.land/manual/contributing

1. Give the PR a descriptive title.

  Examples of good title:
    - fix(std/http): Fix race condition in server
    - docs(console): Update docstrings
    - feat(doc): Handle nested reexports

  Examples of bad title:
    - fix #7123
    - update docs
    - fix bugs

2. Ensure there is a related issue and it is referenced in the PR text.
3. Ensure there are tests that cover the changes.
4. Ensure `cargo test` passes.
5. Ensure `./tools/format.js` passes without changing files.
6. Ensure `./tools/lint.js` passes.
-->

I compared the results of `cli/bench/cache_api.js`.

before

```
cpu: Apple M1
runtime: deno 1.29.1 (aarch64-apple-darwin)

file:///Users/moriken/Developer/deno/cli/bench/cache_api.js
benchmark                  time (avg)             (min … max)       p75       p99      p995
------------------------------------------------------------- -----------------------------
cache_storage_open      29.02 µs/iter  (25.96 µs … 246.21 µs)  29.67 µs  49.33 µs  51.21 µs
cache_storage_has       16.53 µs/iter     (13 µs … 107.71 µs)  17.54 µs  21.08 µs  22.67 µs
cache_storage_delete    21.77 µs/iter  (20.17 µs … 109.17 µs)  21.92 µs  27.75 µs  28.71 µs
cache_put_body_10_KiB     5.1 ms/iter     (2.91 ms … 6.22 ms)   5.72 ms   6.21 ms   6.22 ms
cache_put_no_body          50 µs/iter    (33.75 µs … 8.83 ms)  47.29 µs 133.62 µs 152.42 µs
cache_match             48.67 µs/iter    (42.62 µs … 4.21 ms)  49.62 µs   61.5 µs  68.75 µs
cache_delete            21.26 µs/iter    (16.29 µs … 8.78 ms)  21.71 µs  25.62 µs  27.17 µs
```

after

```
cpu: Apple M1
runtime: deno 1.29.1 (aarch64-apple-darwin)

file:///Users/moriken/Developer/deno/cli/bench/cache_api.js
benchmark                  time (avg)             (min … max)       p75       p99      p995
------------------------------------------------------------- -----------------------------
cache_storage_open      26.46 µs/iter  (21.42 µs … 291.88 µs)  26.21 µs  47.67 µs  48.75 µs
cache_storage_has       15.58 µs/iter    (7.83 µs … 96.29 µs)   16.5 µs  21.62 µs  28.33 µs
cache_storage_delete    20.81 µs/iter  (17.92 µs … 111.25 µs)  22.58 µs  25.38 µs  27.33 µs
cache_put_body_10_KiB    5.02 ms/iter    (1.83 ms … 12.91 ms)   5.21 ms  11.35 ms  12.91 ms
cache_put_no_body       45.49 µs/iter  (28.92 µs … 767.92 µs)  44.25 µs 126.83 µs 148.04 µs
cache_match             46.88 µs/iter     (40.5 µs … 4.26 ms)  48.12 µs  61.79 µs     70 µs
cache_delete            20.77 µs/iter    (16.12 µs … 9.69 ms)  21.71 µs  25.04 µs  26.25 µs
```